pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.60",
    "flask>=3.1.3",
    "python-dotenv>=1.2.2",
    "twilio>=9.10.2",
]
